refactor: mapper package로 제대로 이동 & 의미없는 파일 삭제
