--- src/main/java/com/ODG/ODG_back/dto/participant/EntityMapper.java
@@ -1,6 +1,0 @@
-package com.ODG.ODG_back.dto.participant;
-
-public interface EntityMapper<D, E> {
-    E toEntity(D dto);
-    D toDto(E entity);
-}
