+++ src/main/java/com/ODG/ODG_back/mapper/MeetingCreateResponseMapper.java
@@ -1,4 +1,4 @@
-package com.ODG.ODG_back.dto.participant;
+package com.ODG.ODG_back.mapper;
 
 import com.ODG.ODG_back.domain.Meeting;
 import com.ODG.ODG_back.dto.meeting.response.MeetingCreateResponseDto;
@@ -7,8 +7,7 @@ import org.mapstruct.Mapping;
 import org.mapstruct.Mappings;
 
 @Mapper(componentModel = "spring")
-public interface MeetingCreateResponseMapper extends EntityMapper<MeetingCreateResponseDto, Meeting>{
-    @Override
+public interface MeetingCreateResponseMapper{
     @Mappings({
             @Mapping(source = "inviteCode", target = "linkCode")
     })
