+++ src/main/java/com/ODG/ODG_back/mapper/MeetingInfoResponseMapper.java
@@ -1,4 +1,4 @@
-package com.ODG.ODG_back.dto.participant;
+package com.ODG.ODG_back.mapper;
 
 import com.ODG.ODG_back.domain.Meeting;
 import com.ODG.ODG_back.dto.meeting.response.MeetingInfoResponseDto;
@@ -7,8 +7,7 @@ import org.mapstruct.Mapping;
 import org.mapstruct.Mappings;
 
 @Mapper(componentModel = "spring")
-public interface MeetingInfoResponseMapper extends EntityMapper<MeetingInfoResponseDto, Meeting> {
-    @Override
+public interface MeetingInfoResponseMapper{
     @Mappings({
             @Mapping(source = "title", target = "name"),
             @Mapping(source = "type", target = "purpose")
