+++ src/main/java/com/ODG/ODG_back/mapper/ParticipantRegisterRequestMapper.java
@@ -1,4 +1,4 @@
-package com.ODG.ODG_back.dto.participant;
+package com.ODG.ODG_back.mapper;
 
 import com.ODG.ODG_back.domain.Participant;
 import com.ODG.ODG_back.dto.participant.request.ParticipantRegisterRequestDto;
@@ -9,8 +9,7 @@ import org.mapstruct.Mappings;
 import java.time.LocalDateTime;
 
 @Mapper(componentModel = "spring", imports = {LocalDateTime.class})
-public interface ParticipantRegisterRequestMapper extends EntityMapper<ParticipantRegisterRequestDto, Participant> {
-    @Override
+public interface ParticipantRegisterRequestMapper{
     @Mappings({
             @Mapping(source = "lat", target = "latitude"),
             @Mapping(source = "lng", target = "longitude"),
