+++ src/main/java/com/ODG/ODG_back/mapper/ParticipantRegisterResponseMapper.java
@@ -1,14 +1,12 @@
-package com.ODG.ODG_back.dto.participant;
+package com.ODG.ODG_back.mapper;
 
 import com.ODG.ODG_back.domain.Participant;
 import com.ODG.ODG_back.dto.participant.response.ParticipantRegisterResponseDto;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-import org.mapstruct.Mappings;
 
 @Mapper(componentModel = "spring")
-public interface ParticipantRegisterResponseMapper extends EntityMapper<ParticipantRegisterResponseDto, Participant> {
-    @Override
+public interface ParticipantRegisterResponseMapper{
     @Mapping(source = "id", target = "participantId")
     ParticipantRegisterResponseDto toDto(Participant entity);
 }
