+++ src/main/java/com/ODG/ODG_back/mapper/ParticipantUpdateMapper.java
@@ -1,4 +1,4 @@
-package com.ODG.ODG_back.dto.participant;
+package com.ODG.ODG_back.mapper;
 
 import com.ODG.ODG_back.domain.Participant;
 import com.ODG.ODG_back.dto.participant.request.ParticipantUpdateRequestDto;
@@ -6,7 +6,7 @@ import org.mapstruct.*;
 
 
 @Mapper(componentModel = "spring")
-public interface ParticipantUpdateMapper extends EntityMapper<ParticipantUpdateRequestDto, Participant>{
+public interface ParticipantUpdateMapper{
     @BeanMapping(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE)
     void updateFromDto(ParticipantUpdateRequestDto dto, @MappingTarget Participant participant);
 }
