+++ src/main/java/com/ODG/ODG_back/service/MeetingServiceImpl.java
@@ -5,8 +5,11 @@ import com.ODG.ODG_back.domain.Participant;
 import com.ODG.ODG_back.dto.meeting.request.MeetingCreateRequestDto;
 import com.ODG.ODG_back.dto.meeting.response.MeetingCreateResponseDto;
 import com.ODG.ODG_back.dto.meeting.response.MeetingInfoResponseDto;
-import com.ODG.ODG_back.dto.participant.*;
 import com.ODG.ODG_back.dto.participant.response.ParticipantListResponseDto;
+import com.ODG.ODG_back.mapper.MeetingCreateRequestMapper;
+import com.ODG.ODG_back.mapper.MeetingCreateResponseMapper;
+import com.ODG.ODG_back.mapper.MeetingInfoResponseMapper;
+import com.ODG.ODG_back.mapper.ParticipantListResponseMapper;
 import com.ODG.ODG_back.repository.MeetingRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
